pyproject.toml
[tool.poetry]
name = "github-toolbox"
version = "0.1.0"
description = "My toolbox for GitHub management"
authors = []
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pygithub = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
